+++ src/frontend/apps/impress/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useClipboard';
 export * from './useCmdK';
 export * from './useDate';
+export * from './useHttpErrorMessages';
 export * from './useKeyboardAction';